refactor: add tests for heading and supporting text
