+++ componentsv2/src/main/java/uk/gov/android/ui/componentsv2/supportingtext/GdsSupportingText.kt
@@ -31,7 +31,7 @@ fun GdsSupportingText(
     )
 }
 
-private class SupportingTextPreviewProvider : PreviewParameterProvider<String> {
+internal class SupportingTextPreviewProvider : PreviewParameterProvider<String> {
     override val values: Sequence<String> = sequenceOf(
         "Short supporting text",
         "Long supporting text - Lorem ipsum dolor sit amet, consectetur adipiscing elit, sed do eiusmod tempor incididunt ut labore et dolore magna aliqua. Ut enim ad minim veniam, quis nostrud exercitation ullamco laboris nisi ut aliquip ex ea commodo consequat.",
